// src/game/generated-community-data.ts
export type CommunityDataIdentifiers = keyof typeof _communityDataByIdentifier;
// ...
const _communityDataByIdentifier = {
  "A←": {
    Blocked: [5, 7, 19, 22, 26],
    Patterns: [
      {
        Present: 24,
        Sword: 2,
        Sword3x2: true,
        ConfirmedFoxes: [6, 12, 32, 35],
        UnconfirmedFoxes: [],
      },
      {
        Present: 14,
        Sword: 2,
        Sword3x2: true,
        ConfirmedFoxes: [0, 6, 12, 32, 35],
        UnconfirmedFoxes: [],
      },
      {
        Present: 28,
        Sword: 2,
        Sword3x2: true,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 27,
        Sword: 2,
        Sword3x2: true,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 24,
        Sword: 8,
        Sword3x2: true,
        ConfirmedFoxes: [2, 11, 13, 17, 32],
        UnconfirmedFoxes: [],
      },
      {
        Present: 27,
        Sword: 8,
        Sword3x2: true,
        ConfirmedFoxes: [2, 11, 13, 17, 23],
        UnconfirmedFoxes: [],
      },
      {
        Present: 28,
        Sword: 8,
        Sword3x2: true,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 24,
        Sword: 9,
        Sword3x2: true,
        ConfirmedFoxes: [0, 8, 14, 21],
        UnconfirmedFoxes: [],
      },
      {
        Present: 27,
        Sword: 9,
        Sword3x2: true,
        ConfirmedFoxes: [0, 8, 14, 21],
        UnconfirmedFoxes: [],
      },
      {
        Present: 28,
        Sword: 9,
        Sword3x2: true,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 27,
        Sword: 2,
        Sword3x2: false,
        ConfirmedFoxes: [6, 12, 32, 35],
        UnconfirmedFoxes: [],
      },
      {
        Present: 10,
        Sword: 2,
        Sword3x2: false,
        ConfirmedFoxes: [6, 12, 32, 35],
        UnconfirmedFoxes: [],
      },
      {
        Present: 24,
        Sword: 2,
        Sword3x2: false,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [18, 23],
      },
      {
        Present: 28,
        Sword: 2,
        Sword3x2: false,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 24,
        Sword: 3,
        Sword3x2: false,
        ConfirmedFoxes: [0, 8, 14, 21],
        UnconfirmedFoxes: [],
      },
      {
        Present: 28,
        Sword: 3,
        Sword3x2: false,
        ConfirmedFoxes: [0, 8, 14, 21],
        UnconfirmedFoxes: [],
      },
      {
        Present: 27,
        Sword: 3,
        Sword3x2: false,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 24,
        Sword: 8,
        Sword3x2: false,
        ConfirmedFoxes: [1, 4, 18, 23],
        UnconfirmedFoxes: [],
      },
      {
        Present: 28,
        Sword: 8,
        Sword3x2: false,
        ConfirmedFoxes: [1, 4, 18, 23],
        UnconfirmedFoxes: [],
      },
      {
        Present: 10,
        Sword: 8,
        Sword3x2: false,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 27,
        Sword: 8,
        Sword3x2: false,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 10,
        Sword: 27,
        Sword3x2: true,
        ConfirmedFoxes: [1, 4, 18, 23],
        UnconfirmedFoxes: [],
      },
      {
        Present: 2,
        Sword: 27,
        Sword3x2: true,
        ConfirmedFoxes: [1, 4, 17, 18, 23],
        UnconfirmedFoxes: [],
      },
      {
        Present: 3,
        Sword: 27,
        Sword3x2: true,
        ConfirmedFoxes: [2, 11, 13, 17],
        UnconfirmedFoxes: [],
      },
      {
        Present: 14,
        Sword: 27,
        Sword3x2: true,
        ConfirmedFoxes: [2, 11, 13, 17],
        UnconfirmedFoxes: [],
      },
      {
        Present: 9,
        Sword: 27,
        Sword3x2: true,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 24,
        Sword: 27,
        Sword3x2: true,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
      {
        Present: 8,
        Sword: 27,
        Sword3x2: true,
        ConfirmedFoxes: [],
        UnconfirmedFoxes: [],
      },
    ],
  },
// ...
} as const;
export const communityDataByIdentifier: Record<
  CommunityDataIdentifiers,
  CommunityDataIdentifierPatterns
> = _communityDataByIdentifier;